Fix issue of FASD not initializing correctly
On one machine, for some reason, the env variables and aliases set by
FASD did not reach the resulting shell session.

I don't know why but for some reason the code the shell autodetection
(`--auto`) feature executes via `eval` did not export the aliases and
env vars. The solution is to skip the autodetection and execute the
`eval` code for ZSH directly.

It's the same initialization code but just with one fewer layer of
`eval`.

Since this skips the autodetection shell initializing should be a little
faster, too.